requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
openai
fpdf2
twilio
elevenlabs
supabase